requirements.txt
dlib==19.24.0
fairseq==0.12.2
flow_vis==0.1
imtools==2.0.1
joblib==1.4.2
# k2==1.24.4.dev20240223+cuda10.2.torch1.10.1
kernels==0.9.0
matplotlib==3.5.2
numpy==1.23.0
opencv_contrib_python==4.8.0.74
opencv_python==4.6.0.66
pandas==1.4.3
Pillow==9.2.0
Pillow==11.3.0
prefetch_generator==1.0.3
PyYAML==6.0.2
PyYAML==6.0.2
scikit_learn==1.1.1
seaborn==0.13.2
sentencepiece==0.2.0
soundfile==0.12.1
surgeon_pytorch==0.0.4
thop==0.1.1.post2209072238
timm==1.0.19
torch==2.2.2+cu121
torchaudio==2.2.2+cu121
torchvision==0.17.2+cu121
transformers==4.49.0.dev0
